% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_connection.R
\name{create_con}
\alias{create_con}
\title{Create database connection}
\usage{
create_con(server, database, port, user, password)
}
\arguments{
\item{server}{The server that hosts the database}

\item{database}{The name of the database that includes hockey-fighter data}

\item{port}{The database port number}

\item{user}{Username for the database}

\item{password}{Password for user}
}
\value{
Database connection object
}
\description{
Returns database connection object
}

