% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fighters.R
\name{query_fighters}
\alias{query_fighters}
\title{Get fighter from database}
\usage{
query_fighters(tablename)
}
\arguments{
\item{tablename}{The table in the database with fighter information}

\item{con}{Connection to Postgres database}
}
\value{
A dataframe with a row for each fighter
}
\description{
Returns a dataframe with fighter data
}

